Remove as_expr() and into_expr() methods from Atom
In practice when caller doesn't know whether it have Atom::Expression or
not it is more convenient to match with Atom::Expression(expr) instead
of calling as_expr(). Returning Option<ExpressionAtom> from these
methods doesn't make sense as it leads to duplicate matching. as_gnd()
method is another story because it not only checks the atom type but
also downcasts it to the given type T.

as_expr() method is useful only when caller is confident that atom
instance is an instance of Atom::Expression type and doesn't want to
execute additional code to check it. In such case it should panic when
atom type is not correct. As having such method in API is not safe, I
removed them until we will see the case they really cover.