Replace variables in an atom from the atomspace before matching
Each time when we instantiate the atom we should replace its variables
by unique one. It guarantees that there is no confusion between
variables with same name but in different contexts. Unit tests which
expect that expression with a variable is returned are fixed. New
function atoms_are_equivalent is introduced. It checks that one atom can
be converted to another by renaming variables.