Add atom tree index for searching in GroundingSpace
IndexTree is a tree structure which keeps value accessible by Atom key.
Atom parts are not exactly matched with the IndexTree keys. The
simplified algorithm is used instead:
 - a symbol can be matched by either equal symbol, or grounded atom
   or variable;
 - a grounded atom can be matched by anything, because grounded atom
   can redefined matching procedure;
 - a variable can be matched by anything
 - an expression can be matched either by equal expression or grounded
   atom or variable.
IndexTree::get() funtion returns an iterator of the matched atoms which
then should be matched precisely. Thus index reduces the search time for
some cases.