Add short parts to the end of expressions matched by wildcard
Each expression is represented in MultiTrieNode by two paths. First is
path for the expression matched by the same expression. Second is for
the expression matched by wildcard. This solution twices number of paths
each time when subexpression is presented in key. Which makes a number
of nodes be equal to 2^<number of expressions>. To reduce number of
nodes we adding special field to keep the link to the node matched by
the wildcard. It allows using the number of nodes equal to number of
NodeKeys to keep the TrieKey.