Add hacky Assert...Op implementation
This implementation calls interpret() function inside directly which is
not a minimal MeTTa way. The proper way is to implement it in MeTTa
using pure interpret() and collapse() function. Less proper way is to
return `(interpret $atom %Undefined% <space>)` from grounded Assert...Op
operation. But current implementation is added to quickly migrate to the
minimal MeTTa interpreter.