Properly handle empty result for the function call and tuple call
In case of function call empty result means the function is not defined
on the passed arguments. Thus the whole branch is removed from the plan.
In case of tuple call empty result means there is no function which
matches this tuple. Thus original tuple is returned as a result.