Return NotReducible instead of Empty when function is not defined
This corresponds to the core functionality implemented in a previous
version of the interpreter. When function definition cannot be found the
atom is returned as is. Two major use-cases are type constructors (which
are usually have no definition only type declaration) and partially
defined functions (which return results on some arguments and stay
non-reducible on others).