Fix collapse to keep variable context which was created before calling it
When collapse is called and then results are filtered, in case when all
results are filtered out after collapse we still need to keep bindings
of the variables which were found before calling collapse. When collapse
results are passed to the superpose we need to merge these results with
bindings inside collapse results.