Cleanup variables processing algorithm for minimal MeTTa interpreter
Add a single place to cleanup out of scope variables. Add variables to
the scope set when stack is going deeper (chain and function
operations). Add "apply and cleanup" method to the Bindings to apply
only bindings which are going to be out of scope and keep bindings in
scope.