Use VariableAtom::from_name method to fix DAS gate behavior
New method is introduced in Python VariableAtom API in order to handle
this particular case because it is not a common case. Usually user
constructs VariableAtom instance from the sole name. In rare testing
case user needs to construct variable from both name and id. And for
specific variable manipulation like DAS gate makes we need to convert
variable into a name and back. Thus separate methods are added to do
this conversion.