Check if function type is applicable to the signature before call
Before this logic got the types of the function and tried to apply each to
the parameter's types. Thus each time effectively evaluating the
function result with the given function type. It leads to the
duplication of the function results when the same function has few
different type signatures (non-deterministic type). This commit changes
behavior. New behavior is filtering list of the function type by
checking if type is applicable to the parameter's types first. Then get
any type from the list and call the function using this type.