Import corelib and stdlib into top module in a proper order
After corelib tokens export procedure is introduced the tokens are added
into the end of the tokenizer queue and not at the beginning as it was
before. Thus it is needed to import stdlib first and corelib second to
return the proper order. Also condition is made more specifical to
prevent importing corelib into stdlib.