Add get_atom_type_v2 which returns more information about types
For each type it returns whether the type is application or it is an
application which container argument error, or it is function or value.
Implement get_atom_type on top of the get_atom_type_v2 information.
Don't return partially defined tuple types. For example for (a c) where
(: a A) return %Undefined% instead of (A %Undefined%), %Undefined%.