Return None when trying to convert grounded atom of inappropriate type
Previous code expected grounded atom is deserialized correctly. It is
not a case when converting serializer inputs the grounded atom which
doesn't have the appropriate type. In such case code raised panic. At
the same time if atom is not grounded at all then None is returned. This
commit is to return None when atom is grounded but cannot be
deserialized by specified converting serializer.