Separate corelib tokens which require additional context
There two kind of tokens in corelib (or Rust stdlib):
(1) Token which is instantiated using context of the importing space.
For example assertEquals contains reference to the importing module's
space which is used to interpret passed atoms.
(2) Token which is instantiated without context. It is a more common
case when token is a function or type constructor which state doesn't
depend on a module which imports the token.

This commit separates these kinds of tokens by adding
register_context_dependent_tokens() and
register_context_independent_tokens() functions to register tokens of
the type (1) and type (2) respectively.