Run DAS integration test on Ubuntu only
At the moment it is not possible to run Docker Buildx (which is required
by das-cli) on macos- and windows- GitHub actions images. On the other
hand testing DAS integration comprehensively is not a task of
hyperon-experimental repo. It is suggested to keep only job to make a
smoketest on linux. Be able run this job manually and as a part of
nightly run.